working on new configuration generator file writer
